requirements.txt
python-telegram-bot
Flask
gunicorn
asyncio # Often implicitly installed, but good to include if using async Flask